feat: separated the websocket base class
